Revert "Filter alerts (#147)"
This reverts commit 468a5df48aa84da8eb4cf029b52c3bedd19ccf58.
