+++ api/error.go
@@ -5,10 +5,9 @@ import "fmt"
 type EntityType string
 
 const (
-	EntityTypeParser      EntityType = "parser"
-	EntityTypeAction      EntityType = "action"
-	EntityTypeAlert       EntityType = "alert"
-	EntityTypeFilterAlert EntityType = "filter-alert"
+	EntityTypeParser EntityType = "parser"
+	EntityTypeAction EntityType = "action"
+	EntityTypeAlert  EntityType = "alert"
 )
 
 func (e EntityType) String() string {
@@ -49,13 +48,6 @@ func ActionNotFound(name string) error {
 func AlertNotFound(name string) error {
 	return EntityNotFound{
 		entityType: EntityTypeAlert,
-		key:        name,
-	}
-}
-
-func FilterAlertNotFound(name string) error {
-	return EntityNotFound{
-		entityType: EntityTypeFilterAlert,
 		key:        name,
 	}
 }
