--- api/internal/humiographql/scalars.go
@@ -1,5 +1,0 @@
-package humiographql
-
-type RepoOrViewName string
-
-type Long int64
